+++ CHANGELOG.md
@@ -5,6 +5,12 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [8.30.0]
+
+### Added
+
+- server: add `Devices` field to server details
+
 ## [8.29.2]
 
 ### Fixed
@@ -675,7 +681,8 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 First stable release
 
-[Unreleased]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.29.2...HEAD
+[Unreleased]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.30.0...HEAD
+[8.30.0]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.29.2...v8.30.0
 [8.29.2]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.29.1...v8.29.2
 [8.29.1]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.29.0...v8.29.1
 [8.29.0]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.28.0...v8.29.0